fix: Use `uint32` for cutoff, as elsewhere
For reference the max size of the data types
|:---------|---------------------------:|
| `uint32` |              4 294 967 295 |
| `uint64` | 18 446 744 073 709 551 615 |

Typically, the cutoff value will not go much higher than 10 000.
At that point even the CUDA implementation struggles (~3 spf)

+++ src/mandelbrot/implementations/nvidia_cuda/__init__.py
@@ -84,7 +84,7 @@ class MandelbrotComputer(MandelbrotComputerInterface):
             y_range[1],
             cutoff,
             divergence.data.ptr,
-            cp.uint64(size),
+            cp.uint32(size),
         )
         self.stream.sync()
         return divergence
